Implementing router navigation mobile
